# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agenda+ API",
        "description": "API RESTful para sistema de gestão de consultas médicas Agenda+",
        "contact": {
            "name": "Suporte Agenda+",
            "email": "suporte@agendaplus.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Servidor da API"
        }
    ],
    "paths": {
...
        "/appointments/{id}/observations": {
            "post": {
                "tags": [
                    "Observações"
                ],
                "summary": "Registrar observação clínica",
                "description": "Registra uma observação clínica para uma consulta. Apenas o médico responsável pela consulta pode registrar observações.",
                "operationId": "334aba87c739ce7f7885de2c19d5139f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da consulta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "anamnesis"
                                ],
                                "properties": {
                                    "anamnesis": {
                                        "type": "string",
                                        "example": "Paciente relata dor de cabeça há 3 dias"
                                    },
                                    "diagnosis": {
                                        "type": "string",
                                        "example": "Cefaleia tensional"
                                    },
                                    "prescription": {
                                        "type": "string",
                                        "example": "Paracetamol 500mg, 1 comprimido a cada 8h"
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Retorno em 7 dias se persistir"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Observação registrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Apenas o médico responsável pode registrar observações"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/patient/observations": {
            "get": {
                "tags": [
                    "Observações"
                ],
                "summary": "Listar observações do paciente",
                "description": "Lista observações clínicas do paciente autenticado. Pacientes não veem conteúdo completo (anamnesis, diagnosis, prescription, notes são ocultados por sigilo médico).",
                "operationId": "37640e48f6146057578b5d3becbebb7e",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de observações (conteúdo clínico oculto para pacientes)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/doctor/patients/{id}/observations": {
            "get": {
                "tags": [
                    "Observações"
                ],
                "summary": "Histórico de observações do paciente (médico)",
                "description": "Lista histórico completo de observações de um paciente. Apenas médicos que já atenderam o paciente podem acessar o histórico completo.",
                "operationId": "10b7279cf23a381126348ee3fc936301",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do paciente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Histórico completo de observações",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Médico não possui atendimentos para este paciente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Sanctum",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}